+++ contracts-sdk/src/addresses/ExampleGatedNFTMinterAddress_amoy_staging.ts
@@ -1,2 +1,0 @@
-export const ExampleGatedNFTMinterAddress_amoy_staging =
-  "0x78B1c49Bb4d0D0aAc211710f05596B5e40543483" as const;