+++ contracts-sdk/src/addresses/ExampleGatedNFTMinterAddress_polygonAmoy.ts
@@ -1,0 +1,2 @@
+export const ExampleGatedNFTMinterAddress_polygonAmoy =
+  "0xb56dCa60d505a8284aE9D7aA9012091d7b9E74C0" as const;