+++ contracts-sdk/src/addresses/ExampleGatedNFTMinterAddress_sepolia_prod.ts
@@ -1,2 +1,0 @@
-export const ExampleGatedNFTMinterAddress_sepolia_prod =
-  "0xC56e619A07ecCCf8221Ba2Fe2EBF0496534B9195" as const;