+++ contracts-sdk/src/addresses/ExampleNFTMinterAddress_amoy_prod.ts
@@ -1,2 +1,0 @@
-export const ExampleNFTMinterAddress_amoy_prod =
-  "0x2fb5C2eE1527c89D35Df7D0B7bA9128afe0162AB" as const;