+++ contracts-sdk/src/addresses/ExampleNFTMinterAddress_polygon.ts
@@ -1,0 +1,2 @@
+export const ExampleNFTMinterAddress_polygon =
+  "0xbADd59796A0b3a0e076a4247Ff9fbB63806B9b42" as const;