+++ contracts-sdk/src/addresses/ExampleNFTMinterAddress_polygonAmoy.ts
@@ -1,0 +1,2 @@
+export const ExampleNFTMinterAddress_polygonAmoy =
+  "0x532187e21fb9c24d61109b65f6d10C8CF0747f54" as const;