+++ contracts-sdk/src/addresses/ExampleNFTMinterAddress_sepolia.ts
@@ -1,0 +1,2 @@
+export const ExampleNFTMinterAddress_sepolia =
+  "0x604626b7A71d16Dd929E6222316aD0667984183e" as const;