+++ contracts-sdk/src/addresses/ExampleNFTMinterAddress_sepolia_dev.ts
@@ -1,2 +1,0 @@
-export const ExampleNFTMinterAddress_sepolia_dev =
-  "0x648313dF50207888a14108dfF73F5904751187df" as const;