+++ contracts-sdk/src/addresses/ExampleNFTMinterAddress_sepolia_staging.ts
@@ -1,2 +1,0 @@
-export const ExampleNFTMinterAddress_sepolia_staging =
-  "0x64980c4eb6e38d831093d043CaddC47B7ebb710a" as const;