+++ contracts-sdk/src/addresses/NexeraIDSignerManagerAddress_polygonAmoy.ts
@@ -1,0 +1,2 @@
+export const NexeraIDSignerManagerAddress_polygonAmoy =
+  "0x29A75f22AC9A7303Abb86ce521Bb44C4C69028A0" as const;