+++ sig-gating-contracts/hardhat.config.ts
@@ -2,15 +2,14 @@ import { HardhatUserConfig } from "hardhat/config";
 
 import "@openzeppelin/hardhat-upgrades";
 import "@nomiclabs/hardhat-ethers";
-import "@nomiclabs/hardhat-etherscan";
 import "@nomiclabs/hardhat-waffle";
 import "@nomicfoundation/hardhat-viem";
 import "@typechain/hardhat";
 import "hardhat-gas-reporter";
 import "solidity-coverage";
-import "@openzeppelin/hardhat-upgrades";
 import "hardhat-deploy";
 import "hardhat-deploy-ethers";
+import "@nomicfoundation/hardhat-verify";
 
 import dotenv from "dotenv";
 
@@ -35,13 +34,25 @@ const config: HardhatUserConfig = {
   },
   etherscan: {
     apiKey: {
-      polygonAmoy: `${process.env.ETHERSCAN_AMOY_API_KEY}`,
+      polygon: `${process.env.ETHERSCAN_POLYGON_MAINNET_API_KEY}`,
+      polygonAmoy: `${process.env.ETHERSCAN_POLYGON_MAINNET_API_KEY}`,
+      base: `${process.env.ETHERSCAN_BASE_API_KEY}`,
+      bsc: `${process.env.ETHERSCAN_BSC_API_KEY}`,
+      mainnet: `${process.env.ETHERSCAN_ETHEREUM_API_KEY}`,
+      sepolia: `${process.env.ETHERSCAN_ETHEREUM_API_KEY}`,
+      optimisticEthereum: `${process.env.ETHERSCAN_OPTIMISM_API_KEY}`,
     },
+  },
+  sourcify: {
+    // Disabled by default
+    // Doesn't need an API key
+    enabled: true,
   },
   networks,
   namedAccounts,
   typechain: {
     outDir: "typechain",
+    target: "ethers-v5",
   },
   mocha: {
     timeout: 3000000,